// src/types/data/dailyRecordData.type.ts
type DailyRecordData = {
  daily_record_id: number;
  character_id: number;
  character_solved: number;
  weapon_id: number;
  weapon_solved: number;
  talent_id: number;
  talent_solved: number;
  constellation_id: number;
  constellation_solved: number;
  food_id: number;
  food_solved: number;
};

export default DailyRecordData;